fix(#4590): Fix dependency on integration-runtime-tests in odata
* Make all test dependencies in odata active if basepom.check.skip-all
  is not defined or false, ie. if flash profile active.
