+++ app/connector/odata/pom.xml
@@ -44,12 +44,6 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>io.syndesis.integration</groupId>
-        <artifactId>integration-runtime</artifactId>
-        <version>${project.version}</version>
-        <classifier>tests</classifier>
-      </dependency>
-      <dependency>
         <groupId>org.apache.httpcomponents</groupId>
         <artifactId>httpclient</artifactId>
         <version>4.5.3</version>
@@ -230,12 +224,6 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>io.syndesis.integration</groupId>
-      <artifactId>integration-runtime</artifactId>
-      <classifier>tests</classifier>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
       <groupId>com.github.spotbugs</groupId>
       <artifactId>spotbugs-annotations</artifactId>
       <scope>test</scope>
@@ -377,7 +365,7 @@
           </execution>
         </executions>
       </plugin>
-       <plugin>
+      <plugin>
         <artifactId>maven-dependency-plugin</artifactId>
         <executions>
           <execution>
